Separated color_shape_detectore class into 2 separated classes: color_detector and shape_detector, deleted test.py
