+++ shapefinder/logdata.csv
@@ -68,3 +68,28 @@
 02/10/2023 13:05:56,,CIRCLE
 02/10/2023 13:05:56,,Square
 02/10/2023 13:05:56,,Rectangle
+02/10/2023 14:04:12,,Square
+02/10/2023 14:04:12,,Triangle
+02/10/2023 14:04:12,,CIRCLE
+02/10/2023 14:04:12,,Square
+02/10/2023 14:04:12,,Rectangle
+02/10/2023 14:05:57,,Square
+02/10/2023 14:05:57,,Triangle
+02/10/2023 14:05:57,,CIRCLE
+02/10/2023 14:05:57,,Square
+02/10/2023 14:05:57,,Rectangle
+02/10/2023 14:22:10,,Square
+02/10/2023 14:22:10,,Triangle
+02/10/2023 14:22:10,,CIRCLE
+02/10/2023 14:22:10,,Square
+02/10/2023 14:22:10,,Rectangle
+02/10/2023 14:27:14,,Square
+02/10/2023 14:27:14,,Triangle
+02/10/2023 14:27:14,,CIRCLE
+02/10/2023 14:27:14,,Square
+02/10/2023 14:27:14,,Rectangle
+02/10/2023 14:29:43,,Square
+02/10/2023 14:29:43,,Triangle
+02/10/2023 14:29:43,,CIRCLE
+02/10/2023 14:29:43,,Square
+02/10/2023 14:29:43,,Rectangle
